+++ rust-doomgeneric/src/w_checksum.rs
@@ -14,7 +14,7 @@ extern "C" {
 }
 static mut open_wadfiles: *mut *mut wad_file_t = ::core::ptr::null::<*mut wad_file_t>()
     as *mut *mut wad_file_t;
-static mut num_open_wadfiles: i32 = 0 as i32;
+static mut num_open_wadfiles: i32 = 0;
 unsafe fn GetFileNumber(mut handle: *mut wad_file_t) -> i32 {
     let mut i: i32 = 0;
     let mut result: i32 = 0;
